+++ src/Sharpen.Engine/CSharpFeatures/AsyncAwait.cs
@@ -11,6 +11,8 @@ namespace Sharpen.Engine.CSharpFeatures
 
         public string FriendlyName { get; } = "Async and await";
 
+        public string LearnMoreUrl => "https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/concepts/async/";
+
         public static readonly AsyncAwait Instance = new AsyncAwait();
     }
 }