+++ src/Sharpen.Engine/CSharpFeatures/DefaultExpressions.cs
@@ -11,6 +11,8 @@ namespace Sharpen.Engine.CSharpFeatures
 
         public string FriendlyName { get; } = "Default expressions";
 
+        public string LearnMoreUrl => "https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/operators/default#default-literal";
+
         public static readonly DefaultExpressions Instance = new DefaultExpressions();
     }
 }