+++ src/Sharpen.Engine/CSharpFeatures/ImplicitlyTypedLocalVariables.cs
@@ -11,6 +11,8 @@ namespace Sharpen.Engine.CSharpFeatures
 
         public string FriendlyName { get; } = "Implicitly typed local variables";
 
+        public string LearnMoreUrl => "https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/classes-and-structs/implicitly-typed-local-variables";
+
         public static readonly ImplicitlyTypedLocalVariables Instance = new ImplicitlyTypedLocalVariables();
     }
 }