+++ src/Sharpen.Engine/CSharpFeatures/NameofExpressions.cs
@@ -11,6 +11,8 @@ namespace Sharpen.Engine.CSharpFeatures
 
         public string FriendlyName { get; } = "Nameof expressions";
 
+        public string LearnMoreUrl => "https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/operators/nameof";
+
         public static readonly NameofExpressions Instance = new NameofExpressions();
     }
 }