+++ src/Sharpen.Engine/CSharpFeatures/SwitchExpressions.cs
@@ -11,6 +11,8 @@ namespace Sharpen.Engine.CSharpFeatures
 
         public string FriendlyName { get; } = "Switch expressions";
 
+        public string LearnMoreUrl => "https://docs.microsoft.com/en-us/dotnet/csharp/whats-new/csharp-8#switch-expressions";
+
         public static readonly SwitchExpressions Instance = new SwitchExpressions();
     }
 }