+++ src/Sharpen.VisualStudioExtension/ToolWindows/AnalysisResultTreeViewItems/CSharpVersionTreeViewItem.cs
@@ -7,7 +7,7 @@ namespace Sharpen.VisualStudioExtension.ToolWindows.AnalysisResultTreeViewItems
         public string CSharpVersion { get; }
 
         public CSharpVersionTreeViewItem(BaseTreeViewItem parent, IAnalysisResultTreeViewBuilder treeViewBuilder, int numberOfItems, string csharpVersion)
-            : base(parent, treeViewBuilder, numberOfItems, "C# " + csharpVersion) // TODO: Replace the string creation once the abstraction for the C# version is fully implemented.
+            : base(parent, treeViewBuilder, numberOfItems, "C# " + csharpVersion, null) // TODO: Replace the string creation once the abstraction for the C# version is fully implemented.
         {
             CSharpVersion = csharpVersion;
         }