+++ src/Sharpen.VisualStudioExtension/ToolWindows/AnalysisResultTreeViewItems/SuggestionTreeViewItem.cs
@@ -8,7 +8,7 @@ namespace Sharpen.VisualStudioExtension.ToolWindows.AnalysisResultTreeViewItems
         public ISharpenSuggestion Suggestion { get; }
 
         public SuggestionTreeViewItem(BaseTreeViewItem parent, IAnalysisResultTreeViewBuilder treeViewBuilder, int numberOfItems, ISharpenSuggestion suggestion)
-            : base(parent, treeViewBuilder, numberOfItems, suggestion.FriendlyName)
+            : base(parent, treeViewBuilder, numberOfItems, suggestion.FriendlyName, null)
         {
             Suggestion = suggestion;
         }